Release 5.8.0
* New samples
  * GLES2.System.DevBasicRender
  * GLES2.UI.Benchmark
  * GLES3.System.DevBasicRender
  * GLES3.UI.Benchmark
  * Vulkan.GpuTimestamp
  * Vulkan.System.DevBasicRender
  * Vulkan.UI.Benchmark
* UI rendering
  * Rewrote the UI rendering code to depend on the new FslGraphics3D.BasicRender.
  * Significant performance improvements.
  * Rendering code support various rendering optimizations that can be toggled at run-time.
  * Updated the custom UI guidelines.
* Android
  * Android apps now support command line arguments. Add the arguments using ```-e "args" "<arguments>" for example ```-e "args" "--DensityDpi 160 -vvvv"```
* Dependencies:
  * Upgraded to OpenCV 4.5.2

+++ DemoFramework/FslDemoApp/Base/include/FslDemoApp/Base/Overlay/DemoAppProfilerGraph.hpp
@@ -45,17 +45,17 @@ namespace Fsl
 
   class DemoAppProfilerGraph
   {
-    int32_t m_minValue;
-    int32_t m_maxValue;
+    int64_t m_minValue;
+    int64_t m_maxValue;
     Point2 m_size;
-    CircularFixedSizeBuffer<int32_t> m_entries;
+    CircularFixedSizeBuffer<int64_t> m_entries;
     std::vector<Vector2> m_coords;
 
   public:
-    DemoAppProfilerGraph(const int32_t minValue, const int32_t maxValue, const Point2& size);
+    DemoAppProfilerGraph(const int64_t minValue, const int64_t maxValue, const Point2& size);
 
     Point2 GetSize() const;
-    void Add(const int32_t latestValue);
+    void Add(const int64_t latestValue);
     void Draw(const std::shared_ptr<IBasic2D>& basic2D, const Vector2& dstPosition, const Color& color);
   };
 }
